pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plaud-cli"
version = "0.1.0"
description = "Unofficial command-line tool for the Plaud.ai API"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.0",
    "python-dateutil>=2.9",
    "pyyaml>=6.0",
]

[project.scripts]
plaud = "plaud_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}
